# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger 開放應用程式介面規範",
        "description": "Swagger 開放應用程式介面規範",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "應用程式介面伺服器"
        }
    ],
    "paths": {
        "/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "取得專案清單",
                "description": "回傳專案清單",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "成功的操作",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未經認證"
                    },
                    "403": {
                        "description": "禁止存取"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "儲存新專案",
                "description": "回傳專案資料",
                "operationId": "storeProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "成功的操作",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "錯誤請求"
                    },
                    "401": {
                        "description": "未經認證"
                    },
                    "403": {
                        "description": "禁止存取"
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "取得專案資訊",
                "description": "回傳專案資料",
                "operationId": "getProjectById",
                "parameters": [
                    {
                        "name": "編號",
                        "in": "path",
                        "description": "專案編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功的操作",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "錯誤請求"
                    },
                    "401": {
                        "description": "未經認證"
                    },
                    "403": {
                        "description": "禁止存取"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "更新既有專案",
                "description": "回傳更新專案資料",
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "專案編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "成功的操作",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "錯誤請求"
                    },
                    "401": {
                        "description": "未經認證"
                    },
                    "403": {
                        "description": "禁止存取"
                    },
                    "404": {
                        "description": "找不到資源"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "刪除既有專案",
                "description": "刪除一筆資料且回傳空的內容",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "專案編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "成功的操作",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "未經認證"
                    },
                    "403": {
                        "description": "禁止存取"
                    },
                    "404": {
                        "description": "找不到資源"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Project": {
                "title": "專案",
                "description": "專案模型",
                "properties": {
                    "id": {
                        "title": "編號",
                        "description": "編號",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "名稱",
                        "description": "新專案名稱",
                        "type": "string",
                        "example": "一個好的專案"
                    },
                    "description": {
                        "title": "內容",
                        "description": "新專案內容",
                        "type": "string",
                        "example": "這是新專案內容"
                    },
                    "created_at": {
                        "title": "建立時間",
                        "description": "建立時間",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "更新時間",
                        "description": "更新時間",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "刪除時間",
                        "description": "刪除時間",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "author_id": {
                        "title": "作者編號",
                        "description": "新專案作者編號",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "author": {
                        "title": "作者",
                        "description": "專案作者使用者模型",
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Project"
                }
            },
            "User": {
                "title": "使用者",
                "description": "使用者模型",
                "properties": {
                    "id": {
                        "title": "編號",
                        "description": "編號",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "名稱",
                        "description": "名稱",
                        "type": "string"
                    },
                    "email": {
                        "title": "電子郵件",
                        "description": "電子郵件",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "title": "電子郵件驗證時間",
                        "description": "電子郵件驗證時間",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "created_at": {
                        "title": "建立時間",
                        "description": "建立時間",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "更新時間",
                        "description": "更新時間",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "刪除時間",
                        "description": "刪除時間",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object"
            },
            "ProjectResource": {
                "title": "專案資源",
                "description": "專案資源",
                "properties": {
                    "data": {
                        "title": "資料",
                        "description": "資料封裝技術",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProjectResource"
                }
            },
            "StoreProjectRequest": {
                "title": "儲存專案請求",
                "description": "儲存專案請求內文資料",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "名稱",
                        "description": "新專案名稱",
                        "type": "string",
                        "example": "一個好的專案"
                    },
                    "description": {
                        "title": "內容",
                        "description": "新專案內容",
                        "type": "string",
                        "example": "這是新專案內容"
                    },
                    "author_id": {
                        "title": "作者編號",
                        "description": "新專案作者編號",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateProjectRequest": {
                "title": "更新專案請求",
                "description": "更新專案請求內文資料",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "名稱",
                        "description": "新專案名稱",
                        "type": "string",
                        "example": "一個好的專案"
                    },
                    "description": {
                        "title": "內容",
                        "description": "新專案內容",
                        "type": "string",
                        "example": "這是新專案內容"
                    },
                    "author_id": {
                        "title": "作者編號",
                        "description": "新專案作者編號",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "專案應用程式介面終端"
        }
    ]
}